Rename solutionData to solution
