+++ apps/nextjs-website/src/lib/types/solution.ts
